@@ -20,7 +20,7 @@
         "bin/zf-dependency-injection-cache-warmup"
     ],
     "require": {
-        "php": ">=7.3 | >=8.0",
+        "php": "~7.4.0 || ~8.0.0 || ~8.1.0",
         "laminas/laminas-servicemanager": "^2.7 | ^3.0",
         "laminas/laminas-modulemanager": "^2.7",
         "psr/simple-cache": "^1.0"
@@ -50,15 +50,16 @@
         "phpunit/phpunit": "^9.4",
         "doctrine/orm": "^2.5",
         "php-coveralls/php-coveralls": "^2.0",
-        "symfony/yaml": "^4",
-        "friendsofphp/php-cs-fixer": "^2.10",
-        "phpstan/phpstan": "^0.12.0",
+        "symfony/yaml": "^4.0 | ^5.0 | ^6.0",
+        "friendsofphp/php-cs-fixer": "^3.4",
+        "phpstan/phpstan": "^1.3",
         "phpspec/prophecy-phpunit": "^2.0",
         "laminas/laminas-hydrator": "^4.2",
         "laminas/laminas-validator": "^2.14",
         "laminas/laminas-filter": "^2.11",
         "laminas/laminas-form": "^3.0",
-        "laminas/laminas-inputfilter": "^2.12"
+        "laminas/laminas-inputfilter": "^2.12",
+        "bnf/phpstan-psr-container": "^1.0"
     },
     "extra": {
         "zf": {
