@@ -15,10 +15,7 @@ use Reinfi\DependencyInjection\Injection\InjectionInterface;
  */
 class ContainerResolver implements ResolverInterface
 {
-    /**
-     * @var ContainerInterface
-     */
-    private $container;
+    private ContainerInterface $container;
 
     public function __construct(ContainerInterface $container)
     {
