@@ -5,14 +5,9 @@ declare(strict_types=1);
 namespace Reinfi\DependencyInjection\Config;
 
 /**
- * Interface ModuleConfig
- *
  * @package Reinfi\DependencyInjection\Config
  */
 interface ModuleConfig
 {
-    /**
-     * @var string
-     */
-    const CONFIG_KEY = 'reinfi.dependencyInjection';
+    public const CONFIG_KEY = 'reinfi.dependencyInjection';
 }
