@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Reinfi\DependencyInjection\Exception;
 
+use Exception;
+use ReflectionClass;
 use ReflectionParameter;
 
 /**
@@ -11,30 +13,9 @@ use ReflectionParameter;
  *
  * @package Reinfi\DependencyInjection\Exception
  */
-class AutoWiringNotPossibleException extends \Exception
+class AutoWiringNotPossibleException extends Exception
 {
-    /**
-     * @param ReflectionParameter $reflParameter
-     *
-     * @return AutoWiringNotPossibleException
-     */
-    public static function fromParameterName(ReflectionParameter $reflParameter): self
-    {
-        return new self(
-            sprintf(
-                'Could not resolve parameter %s to to a class',
-                $reflParameter->getName()
-            )
-        );
-    }
-
-    /**
-     * @param string $className
-     * @param \ReflectionClass|null $constructedClass
-     *
-     * @return AutoWiringNotPossibleException
-     */
-    public static function fromClassName(string $className, ?\ReflectionClass $constructedClass): self
+    public static function fromClassName(string $className, ?ReflectionClass $constructedClass): self
     {
         return new self(
             sprintf(
@@ -45,11 +26,6 @@ class AutoWiringNotPossibleException extends \Exception
         );
     }
 
-    /**
-     * @param ReflectionParameter $reflParameter
-     *
-     * @return AutoWiringNotPossibleException
-     */
     public static function fromMissingTypeHint(
         ReflectionParameter $reflParameter
     ): self {
@@ -64,11 +40,6 @@ class AutoWiringNotPossibleException extends \Exception
         );
     }
 
-    /**
-     * @param ReflectionParameter $reflParameter
-     *
-     * @return AutoWiringNotPossibleException
-     */
     public static function fromBuildInType(ReflectionParameter $reflParameter): self
     {
         $declaringClass = $reflParameter->getDeclaringClass();
