@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace Reinfi\DependencyInjection\Exception;
 
+use Exception;
+
 /**
  * @package Reinfi\DependencyInjection\Exception
  */
-class ConfigPathNotFoundException extends \Exception
+class ConfigPathNotFoundException extends Exception
 {
 }
